Version 0.0.2 with major API changes
Add top-level get_resolver() function that does most of the work of
locating and setting up resolvers automatically.  Resolvers found in
carmen.resolvers are automatically imported and used when possible.

Ship a built-in locations database, and use it when adding locations to
the resolver if no alternative is explicitly specified.

Don't raise needless warnings when adding locations to resolvers.