Construct Carmen>Geonames Mapping for State/County
Currently, there are still issues since we are not checking aliases or alternative names.
To fix this, we can begin iterating through the aliases when trying to construct the
mapping.

The current algorithm works by querying our database for the Geoname IDs that correspond
with particular country and state names that we are looking for.