locations.json -> new.json, First Version
This is the first attempt at mapping locations.json to a new file,
called new.json. The new JSON representation of the locations
should contain all of the same information as that of locations.json,
with the exception that the ID (id + parent_id) fields are swapped
to use the Geonames IDs.

Note that because this is the first version, there are still issues
and not all of the locations were able to be mapped properly. Only
about 65% of the locations are mapped.