JSON Mapping Uses Geonames Parent IDs
The mapping now uses Geonames' hierarchy.txt instead of the original
locations.json's parent_id field when mapping. That is, the parent_id
field is now ignored in the mapping, and the actual parents are
instead looked up in hierarchy.txt.

Another small update was updating the
carmen/mappings/regions_errors.txt file.