Write Script to Compute Evaluation Metrics
carmen/scripts/get_evaluation_metrics.py computes the accuracy of
our resolver for cities, counties, states, and countries. We can run
it by providing two arguments:

- true_locations.json

- predicted_locations.json

where true_locations.json represents a bunch of locations resolved
using the the geotag resolver--this represents as much of truth as
we can get. predicted_locations.json represents a bunch of locations
resolved using the profile resolver, which is what we want to
evaluate.